app/requirements.txt
aiohttp==3.8.4
aiohttp-socks==0.7.1
aiofiles==0.8.0
aioredis==2.0.1
asyncio==3.4.3
beautifulsoup4==4.10.0
cryptography==3.4.7
fake-useragent==0.1.11
js2py==0.71
lxml==4.6.3
pyppeteer==0.2.6
python-dotenv==0.19.0
requests==2.26.0
stem==1.8.0
tenacity==8.0.1
tldextract==3.1.2
2captcha-python==1.1.3
tensorflow==2.6.0
scikit-learn==0.24.2
numpy==1.21.2
opencv-python==4.5.3.56
esprima==4.0.1
jsbeautifier==1.14.0
transformers==4.11.3
torch==1.9.0
spacy==3.1.3
flask==2.0.1
flask-socketio==5.1.1